# 8. Application/assets/swagger.json
{
    "definitions": {},
    "info": {
        "title": "API REST documentation",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "description": "API REST pour gérer des utilisateurs"
    },
    "basePath": "/api/v1",
    "host": "localhost:8080",
    "tags": [
        {
            "name": "members",
            "description": "Méthodes permettant de gérer les utilisateurs"
        }
    ],
    "paths": {
        "/members/{id}": {
            "get": {
                "summary": "/members/{id}",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {},
                "tags": [
                    "members"
                ]
            },
            "put": {
                "summary": "/members/{id}",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {},
                "tags": [
                    "members"
                ]
            },
            "delete": {
                "summary": "/members/{id}",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {},
                "tags": [
                    "members"
                ]
            }
        },
        "/members": {
            "get": {
                "summary": "/members",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "max",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {},
                "tags": [
                    "members"
                ]
            },
            "post": {
                "summary": "/members",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {},
                "tags": [
                    "members"
                ]
            }
        }
    },
    "swagger": "2.0"
}